move App.js to src folder
